Refactor: 로또 당첨을 계산하는 메서드를 WinningNumbers로 이동
